更新若干代码

+++ VehicleDispatchSystem_MS/vds-gateway/src/main/java/cn/lingbaocrisps/gateway/utils/JwtTool.java
@@ -40,7 +40,7 @@ public class JwtTool {
      * @param token token
      * @return 解析刷新token得到的用户信息
      */
-    public Long parseToken(String token){
+    public Integer parseToken(String token){
         //1.校验token是否为空
         if (token == null)
             throw new UnauthorizedException("未登录");
@@ -64,7 +64,7 @@ public class JwtTool {
         }
         //5.数据解析
         try {
-            return Long.valueOf(userPayload.toString());
+            return Integer.valueOf(userPayload.toString());
         }catch (Exception e){
             throw new UnauthorizedException("无效的token");
         }
